refactor(frontend): refactor tips
