+++ packages/frontend/src/pages/settings/other.vue
@@ -161,8 +161,7 @@ import { prefer } from '@/preferences.js';
 import MkRolePreview from '@/components/MkRolePreview.vue';
 import { signout } from '@/signout.js';
 import { migrateOldSettings } from '@/pref-migrate.js';
-import { store } from '@/store.js';
-import { hideAllTips as _hideAllTips } from '@/store.js';
+import { hideAllTips as _hideAllTips, resetAllTips as _resetAllTips } from '@/tips.js';
 
 const $i = ensureSignin();
 
@@ -206,7 +205,7 @@ function migrate() {
 }
 
 function resetAllTips() {
-	store.set('tips', {});
+	_resetAllTips();
 	os.success();
 }
